feat: 채팅방 헤더 구현
