+++ one-n/src/pages/ChatRoom/sub-components/Footer/Footer.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Footer() {
+    return <div>Footer</div>;
+}
+
+export default Footer;
