+++ one-n/src/pages/ChatRoom/sub-components/Header/Header.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+import MainHeader from "./sub-components/MainHeader/MainHeader";
+import SubHeader from "./sub-components/SubHeader/SubHeader";
+
+function Header() {
+    return (
+        <div>
+            <MainHeader />
+            <SubHeader />
+        </div>
+    );
+}
+
+export default Header;
