+++ one-n/src/pages/ChatRoom/sub-components/Header/sub-components/SubHeader/SubHeader.jsx
@@ -1,0 +1,27 @@
+import React from "react";
+
+import BorderButton from "./sub-components/BorderButton/BorderButton";
+import PopoverButton from "./sub-components/PopoverButton/PopoverButton";
+
+import { ReactComponent as CalendarIcon } from "../../../../../../assets/icons/calendar.svg";
+import { ReactComponent as WonIcon } from "../../../../../../assets/icons/won.svg";
+
+import styles from "./SubHeader.module.scss";
+
+function SubHeader() {
+    return (
+        <div className={styles.SubHeader}>
+            <div className={styles.container}>
+                <div className={styles.leftSection}>
+                    {/* <BorderButton icon={<CalendarIcon />} />
+                    <BorderButton icon={<WonIcon />} /> */}
+                </div>
+                <div className={styles.rightSection}>
+                    <PopoverButton />
+                </div>
+            </div>
+        </div>
+    );
+}
+
+export default SubHeader;
