+++ one-n/src/pages/ChatRoom/sub-components/Header/sub-components/SubHeader/SubHeader.module.scss
@@ -1,0 +1,14 @@
+.SubHeader {
+    padding: 1rem 0;
+
+    .container {
+        padding: 0 1rem;
+        display: flex;
+        justify-content: space-between;
+
+        .leftSection {
+            display: flex;
+            gap: 0.5rem;
+        }
+    }
+}
