+++ one-n/src/pages/ChatRoom/sub-components/Header/sub-components/SubHeader/sub-components/BorderButton/BorderButton.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+import styles from "./BorderButton.module.scss";
+
+function BorderButton({ icon, label, onClick }) {
+    return (
+        <div className={styles.BorderButton}>
+            {icon}
+            <p>{label}</p>
+        </div>
+    );
+}
+
+export default BorderButton;
