+++ one-n/src/pages/ChatRoom/sub-components/Header/sub-components/SubHeader/sub-components/BorderButton/BorderButton.module.scss
@@ -1,0 +1,5 @@
+.BorderButton {
+    padding: 0.5rem 0.75rem;
+    border: 1px solid #eaeaea;
+    border-radius: 8px;
+}
