+++ one-n/src/pages/ChatRoom/sub-components/Header/sub-components/SubHeader/sub-components/PopoverButton/PopoverButton.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function PopoverButton() {
+    return <div>Popover Button</div>;
+}
+
+export default PopoverButton;
