+++ one-n/src/pages/ChatRoom/sub-components/MessageList/MessageList.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function MessageList() {
+    return <div>Message List</div>;
+}
+
+export default MessageList;
